Add C interpreter dispatch using "nested switch"
This is what Andrew wants Zig to lower to based on IRC discussion.

Seems to lower to something equivelant to LLVM indirectbr (but using
only switch instruction).
Unfortunately LLVM duplicates the branch tables, resulting in code
duplication.

Even worse, there is potential for quadratic blowup in the control flow
graph, making optimization algorithms super slow.

Here is a description of LLVM's solution for this for indirectbr:
https://blog.llvm.org/2010/01/address-of-label-and-indirect-branches.html

> By trying to maintain at most one indirectbr instruction per function, we effectively get a factoring of the edges.
> The llvm-gcc and clang frontends both generate IR which has at most one indirectbr instruction per function:
> all other indirect gotos are lowered as a branch to the communal indirectbr in a function.

> In order to produce efficient code, the code generator performs tail duplication to introduce the N^2 CFG

Unfortunately, we don't have this advantage with "nested switch".
We will have a horrible blowup of the CFG and we have now way to force.

So Andrew's design works around sparse jump tables with indirectbr (my concern),
but risks quadratic blowup in optimization passes *sigh*